型変換

+++ TeamWork/MyGame/scr/actor/Field/ClothesLine.cpp
@@ -39,7 +39,7 @@ void ClothesLine::Draw() const
 {
 	Vector2 pPos = world_->GetKeepDatas().playerPos_;
 
-	float sizex = (int)Sprite::GetInstance().GetSize(SPRITE_ID::GAMEPLAY_LANE_SPRITE).x;
+	float sizex = (float)Sprite::GetInstance().GetSize(SPRITE_ID::GAMEPLAY_LANE_SPRITE).x;
 	
 	float drawx = (float)((int)world_->GetInv().Translation().x %(int)sizex);
 	drawx -= 250.0f;
@@ -51,7 +51,7 @@ void ClothesLine::Draw() const
 	Vector2 vecaddpos = Vector2(0, LaneSorigin.y);
 
 	for (int i = 0; i < 6; i++) {
-		Vector2 drawLPos = drawVec + Vector2(sizex / 2.f, 0.f)*i;
+		Vector2 drawLPos = drawVec + Vector2(sizex / 2.f, 0.f)*(float)i;
 		if ((drawEndPos).x <= drawLPos.x + sizex){
 			Vector2 lastdp = drawLPos;// -Vector2(0.0f, Sprite::GetInstance().GetSize(SPRITE_ID::LANE_SPRITE).y / 2);
 			Vector2 lastep = drawEndPos + vecaddpos+Vector2(20.0f, (float)Sprite::GetInstance().GetSize(SPRITE_ID::GAMEPLAY_LANE_SPRITE).y/2.f);
