+++ TeamWork/MyGame/scr/scene/addScreen/GameClearScreen.cpp
@@ -47,7 +47,7 @@ void GameClearScreen::Init()
 	isHeadDraw_ = false;
 	inputCount_ = 0;
 	sinCount_ = defSinC;
-	for (int i = 0; i < changeSceneList_.size(); i++) {
+	for (int i = 0; i < (int)changeSceneList_.size(); i++) {
 		textAlphaList_[i] = 1.f;
 		textSizeList_[i] = 1.f;
 	}
@@ -180,8 +180,8 @@ void GameClearScreen::ScoreUpdate()
 		if (!isKeyOnceScore_) {
 			TweenManager::GetInstance().Cancel(&fscore_);
 			TweenManager::GetInstance().Cancel(&fheadCount_);
-			fscore_ = score_;
-			fheadCount_ = headCount_;
+			fscore_ = (float)score_;
+			fheadCount_ = (float)headCount_;
 			SetFullStarCount();
 			isHeadDraw_ = true;
 			dstar_.SetFull();
@@ -194,8 +194,8 @@ void GameClearScreen::ScoreUpdate()
 		}
 	}
 	if (isKeyOnceScore_) {
-		fscore_ = score_;
-		fheadCount_ = headCount_;
+		fscore_ = (float)score_;
+		fheadCount_ = (float)headCount_;
 
 	}
 	dstar_.Update();
@@ -225,11 +225,11 @@ void GameClearScreen::drawUpdate()
 	sinCount_ += 3;
 	if (sinCount_ > 360)sinCount_ = 0;
 
-	for (int i = 0; i < changeSceneList_.size(); i++) {
+	for (int i = 0; i < (int)changeSceneList_.size(); i++) {
 		textAlphaList_[i] = 1.f;
 		textSizeList_[i] = 1.f;
 	}
-	textAlphaList_[inputCount_] = abs(MathHelper::Sin(sinCount_));
+	textAlphaList_[inputCount_] = abs(MathHelper::Sin((float)sinCount_));
 	textSizeList_[inputCount_] = mxmSize;
 
 }
